Fix du type des Ids et ajout d'annotations @Override manquantes [BCKND]

+++ back-end/src/main/java/com/ecommerce/backend/orders/services/OrderRepository.java
@@ -8,6 +8,6 @@ import com.ecommerce.backend.orders.model.Order;
 
 @Repository
 @RepositoryRestResource(exported = false)
-public interface OrderRepository extends PagingAndSortingRepository<Order, Integer> {
+public interface OrderRepository extends PagingAndSortingRepository<Order, Long> {
 
 }
